refactor: reduced entire search view complexity
The index.tsx related to the search view has been improved greatly by reducing its complexity by extracting many components.
